Tweak tests and stub
