+++ stubs/draft.stub
@@ -1,13 +1,5 @@
 models:
-  Post:
-    title: string:400
-    content: longtext
-    published_at: nullable timestamp
+  # ...
 
 controllers:
-  Post:
-    index:
-      query: all
-      render: post.index with:posts
-
   # ...
